Bound bitvector constants without a negative shift, and spell the 64-bit C API types uint64_t (#705)
* Bound bitvector constants without a negative shift

vc_bvConstExprFromInt checks the caller's value against the largest one
the requested width can hold, computed as 0xFF..FF >> (64 - n_bits). That
shift distance goes negative as soon as n_bits passes 64, which is
undefined; on x86-64 the distance is masked to six bits, so the bound
wrapped back to something small rather than growing. Width 65 admitted
only 0 and 1, width 66 only up to 3, and any larger constant took
FatalError -- the API killing the caller's process over a constant that
fits with room to spare.

Which widths broke depended on the value, so there was no rule to
document around: 70 and 128 pass a constant of 5 only because their
collapsed bounds land at 63 and UINT64_MAX. A bigger value fails there
too.

Compute the bound with a branch instead, and reject a non-positive width
up front rather than leaving it to shift by 64 (also undefined) and then
reach CreateBVConst, which only diagnoses zero. For every width the old
expression was defined for, 1 to 64, the new one gives the identical
bound -- checked exhaustively -- so nothing that was accepted before is
rejected now.

The locals become uint64_t rather than unsigned long long int, which is
what the width arithmetic actually means, and the format strings move to
PRIu64 to match. The failure message now names the width and the maximum
instead of printing "5 >= 1", which read as though 5 were the smaller.

The new test covers widths either side of the word boundaries with the
values that used to abort; it fails on master by aborting the process.

* Spell the 64-bit C API types uint64_t

getBVUnsignedLongLong and vc_bvConstExprFromLL are declared with
unsigned long long int, which says how the type was once spelled rather
than what it means. Both are 64-bit quantities; say so.

The ABI does not move. Both are extern "C", so the exported symbols are
unmangled and unchanged. On i386 and on Windows the edit is a no-op,
because uint64_t is unsigned long long there. On LP64 the spelling
becomes unsigned long, which is the same 64 bits passed the same way --
checked by linking a uint64_t-declared caller against a libstp built
with the old spelling, and an old-spelling caller against a libstp built
with the new one: every boundary value from 0 to UINT64_MAX round-trips
either way.

The Python binding has always declared both as c_uint64, so the header
had been the less precise of the two. It now agrees with the binding it
ships alongside.

stdint.h joins stdbool.h at the top. The header still compiles clean as
C99, C11 and C++17 with -Wall -Wextra -pedantic.

The names stay. Renaming getBVUnsignedLongLong would be a real break for
callers rather than a change of spelling, and it is not worth a
deprecation cycle to lose the "LongLong" in a name that has always meant
64-bit. The one caller this can affect is code taking the address of
either function into an explicitly-typed function pointer: a hard error
in C++, a warning in C. Ordinary calls are unaffected.

+++ include/stp/c_interface.h
@@ -36,6 +36,7 @@ extern "C" {
 #endif
 
 #include <stdbool.h>
+#include <stdint.h>
 
 /////////////////////////////////////////////////////////////////////////////
 /// STP API INTERNAL MACROS FOR LINKING
@@ -572,13 +573,13 @@ DLL_PUBLIC int getBVInt(Expr e);
 //!
 DLL_PUBLIC unsigned int getBVUnsigned(Expr e);
 
-//! Return an unsigned long long int from a constant bitvector expressions
+//! Return a uint64_t from a constant bitvector expressions
 
-//! \brief Returns the associated unsigned long long integer from the given bitvector expression.
+//! \brief Returns the associated 64-bit unsigned integer from the given bitvector expression.
 //!
-//! Panics if the given bitvector cannot be represented by an 'unsigned long long int'.
+//! Panics if the given bitvector cannot be represented by a 'uint64_t'.
 //!
-DLL_PUBLIC unsigned long long int getBVUnsignedLongLong(Expr e);
+DLL_PUBLIC uint64_t getBVUnsignedLongLong(Expr e);
 
 //! \brief Prints the bit string for a a constant bitvector expression to a
 //!        buffer allocated by STP that is returned via output parameter 'buf'
@@ -636,12 +637,11 @@ DLL_PUBLIC Expr vc_bvConstExprFromStr(VC vc, const char* binaryInput);
 DLL_PUBLIC Expr vc_bvConstExprFromInt(VC vc, int bitWidth, unsigned int value);
 
 //! \brief Returns a bitvector with 'bitWidth' bit-width from the given
-//!        unsigned long long integer value.
+//!        64-bit unsigned integer value.
 //!
 //! The 'bitWidth' must be large enough to fully store the given value's bit representation.
 //!
-DLL_PUBLIC Expr vc_bvConstExprFromLL(VC vc, int bitWidth,
-                                     unsigned long long value);
+DLL_PUBLIC Expr vc_bvConstExprFromLL(VC vc, int bitWidth, uint64_t value);
 
 //! \brief Returns a bitvector with a bit-width of 32 from the given
 //!        unsigned integer value.
